Add support for generating per-policy module RPMs.
It is beneficial to generate RPMs that contain
a single policy module so they can be referenced
in a kickstart.  E.g., a normal CLIP build
doesn't run Apache, but the Apache variant does.
So generate separate RPMs for the apache polcy
module.

Use the SEPARATE_PKGS variables
in the policy Makefile to list policy
modules that should end up in a separate
RPM.