append to the FORWARD chain rather than insert at the top
The default _updown script would append a rule that ACCEPT's packets
being forwarded as part of the connection.  This would cause the
rule we've added to reset the mss of forwarded TCP packets to not get
hit, which causes usability issues.

Append to the FORWARD chain instead.  Now the mss will still be updated,
and ipsec traffic won't get dropped