Adjust src.rpm Makefile dep.
exut 1'ing intentionally isn't cool.
The first manifestation of the problem is in cacti on
clean checkouts.  It looks like Makefile.tmpl was
getting checked out after Makefile so on a clean checkout
the first run of make always failed.

Instead, assume a Makefile was committed and only bail
when noon exists at all.

Close #46