Use the complete yum conf for repo queries.
repo queries were broken for top-level RPM
builds, eg, clip-selinux-policy-rpm.  This
is because the queries were done using the
truncated yum conf.  It worked for centos
repos because the yum conf is correct when
they are poopulated, but not true for clip-repo.
Switching to the concatenated yum conf fixes this.

Closes #53