Address some dep problems that broke builds after bare.
make bare bare-repos led to broken builds as it
led to a repos/clip-repo that wasn't a valid yum repo.
In trying to unwind this, I relized we had led ourselves
into a new circle of hell dedicated to PHONIES.  At some
point we must stop or else we will end up in that circle
for eternity.