Grab both CD-size and DVD-size ISOs after rolling.
If a ISO goes over 750M Pungi puts DVD in the name.
Account fot that by using a wildcard when moving
the iSO arounGrab both CD-size and DVD-size ISOs after rolling.

If a ISO goes over 750M Pungi puts DVD in the name.
Account fot that by using a wildcard when moving
the iSO aroundd