Update auditd access to audit logs for RHEL 7.
In v7 auditd fully manages the audit log. Add
the necessary perms.

I'm still torn on this. Just because it is trusted
to add entries doesn't mean it should be trusted to
truncate or unlink. If it not compromised at point A,
then becomes compromised at point B, the compromised
daemon can destrory all records leading to the compromise.