Fix setting of SELinux enforcement mode.
CONFIG_BUILD lets you specify an SELinux
enforcement mode that is supposed to
propagate into /etc/selinux/config on the
deployed system. It wasn't happening.

It appears that, at some point, anaconda
starting modifying the SELINUX= value to
match the setting in the kickstart header.
E.g. if the header had "selinux --enforcing"
/etc/selinux/config always had SELINUX=enforcing
regardless of what we did in %post.

To work around this issue, when we modify the
kickstart during a build, also modify the
selinux field in the header if COMFIG_BUILD
specifies permissive mode.