Fix issue with ctrl-alt-del handling.
We were replacing the ExecStart command
in the reboot target with "logged" trying
to retain similar functionality from RHEL
6 STIGs. But that prevented any version
of "reboot" from working.

We were already masking the ctrl-alt-del
target so I'm not sure how we endded up
with this too. The masked target generates
a log message, albeit in a different log,
so stop using logger.