Allow systemd-readahead to notify systemd.
We haven't split systemd-readahead out into
a separate domain. It tries to notify systemd
but fails to and this can lead to weird things
like disks not being mounted (!?). Furrther,
this popped up in a bare metal install but we
hadn't seen it before in VMs (?).

Add the policy thaty lets it notify systemd.