Strip duplicates from PACKAGES variable.
PACKAGES is partially composed of the
kickstart-specific variants_pkgs.mk.
If multiple kickstarta have the same
package name in variants_pkgs.mk we
end up with dupes in packages and when
we generate recipes from this variable
it ends up redeclaring the same recipe
multiple times.

To avoid this warning strip dupes from the
variable before using it to generate recipes.