Switch SELinux policy to 7.3 layout.
RH switched to new policy layout and
CIL in 7.3 This requires a major
overhaul in how policy is packaged up.

Update SELinux policy build support to
use the new layout.

Note - as of this commit, the policy
no longer works for 7.2.

Backwards compatibility will be add
introduced in a subsequent commit.