Fix detection of which version of policy spec files should be used.
The addition of support for RHEL 7.3 necessitated selecting
policy spec files based on which version of RHEL was targeted.
The previous implmentation failed to fully address this issue
and ended up building the wrong version in certain conditions.

Fix that and also query the previously used spec file to figure out
if the OS version targeted actually changed since the last build
so it will rebuild when a make var changes.