Add top-level ADDTL_DEPS to dep list for RPM builds.
We were leveraging the existing top-level ADDTL_DEPS
variable wen building the dep list for rolling srpms/rpms.

This hasn't led to significant problems, but with the
introduction of separate spec files plus dynamic detction
of OS releases for SELinux policy, changes to the top-level
become very relevant.

I think this has led to bugs in the past too, but they
aren't reliably reproducible because it is highly dependent
on how a user has interacted with the repo.