Fix OS version/release detection.
For hopefully the final time, fix
how we detect the OS version plus release
number.

The final bug was that, after a make bare,
or clean checkout, conf/yum/yum_all.conf did
not exit. So, whenever we tried to access
repos in that file it returned nothing.

The easy fix is to turn it in a make define
we can call from whatever context we want,
including directly when we run submakes for
src RPMs.

Also rename OS_VER to OS_REL to indicate it
contains release as opposed to the two other
OS_VER vars.