Fix relabeling between policy updates.
Relabeling was functional when installing
the 7.3 CIL RPM the first time, then things
broke, and you couldn't login b/c init was
not labaled properly (and a lot of other things).

Overhaul policy labeling in the spec file in
pre and post sections. A lot of this came
directly from RH's spec, but others, such
as autorelabel on reboot, was not present
but for some reason we still need it.