Use semanage man page to determine CIL support.
There is probably a much more elegant way of doing
this, but I'm tired of chasing down these CIL issues.

RH decided to backport CIL to 7.2.x (no further comment).
So all of the work to properly determine the RHEL version
is moot. Screw it, look at the man page for semanage and
be done with it.