Only generate RPMs for one policy type.
Building both MCS and MLS SELinux policies
takes a long time. Most of the time, you only
want one or the other. Add a TYPE variable to
the policy Makefile allowing a use to select
whichg policy type to build.

Note that this only works for CIL-based policies
for now. Non-CIL policies use the older spec file
which has not been update with support for the TYPE
variable.