Enable man policy module.
As a consequence of adding support
for plymouth graphical boot, a huge
amount of font RPMs is pulled in
(a few hundred megs). Thus, man pages
are no longer the long pole in the tent,
and not having them is a development and
admin burden.

Enable the mandb policy module which provides
support for man.