support faster development builds
added targets and scripts to use an existing install iso to re-roll a
new one with updated packages and kickstart.  this must not be used
for production builds because the installation (anaconda) environment
and build/version information does not get updated.
to use, copy an existing install iso to shell.iso and run
make VARIANT-inst-iso-fast