Fix EFI/leggacy BIOS compatibility.
The logic in lorax WRT snagging the correct
boot image was not working for legacy BIOS.
It happened to work for UEFI due to a logic
and a syntax error.

Add a conditional that will snag the correct
boot image.

The modifications in this commit are copyrighted works of Owl Cyber Defense.
(C) 2018 Owl Cyber Defense Solutions, LLC