Use comps.xml from user-specified yum repos.
We have been carrying a comps.xml file and
occasionally updating it with what is in
RHEL proper. That introduces problems since
the the comps.xml in our rrepo may not reflect
what is upstream until we update it.

Instead, use the comps.xml if it is present in
the user-specified repos and fall back to our in-tree
version when one isn't available.

The modifications in this commit are copyrighted works of Owl Cyber Defense.
(C) 2019 Owl Cyber Defense Solutions, LLC