Fix ARCH declarations in package Makefiles
We eval the ARCH variable specified in
package Makefiles. The issue is that if
one uses "?=" in the declaraction it
means the ARCH from one package won't
replace the ARCH from the previous package.
This manifests itself in packages that
continually get built ever without modifications.

This fixes the declaractions so they are all ":=".
This is still fragile.

Alterntive 1: in the eval cleanup the variable value
prior to the eval itself thus replacing whtever is there
with ":=".

Alterntive 2: do more recursive Make stuff letting the
package tell the top-level what the package name is.
I'm trying to avoid more recursion, but this seems like
the cleanest approach (thanks Marshall!).

The modifications in this commit are copyrighted works of Owl Cyber Defense.
(C) 2019 Owl Cyber Defense Solutions, LLC