Split GUI packages into separate include file.
Split the GUI package list into a separate
%include file. This allows users to have
the GUI enabled at the top-level, but disable
it for specific kickstarts.

Add the %include entries for the GUI
to the remaining kickstart files.

The modifications in this commit are copyrighted works of Owl Cyber Defense.
(C) 2019 Owl Cyber Defense Solutions, LLC