Associate toor_u with system_r role.
I'm not sure how we made it this far,
but the toor user needs to be associated
with system_R for basic things to work,
like RPM.

RPM's policy was broken so we didn't
notice rpm running as toor_t until it
was fixed and the problem became apparent.
The problem was due to a role trans from
toor_r to system_r but, while the trans
was in-place, the context was invalid so
nothing worked.