Fixup some SELinux network problems.
First, ping sends a UDP package to port
1025 on the destination host to figure
out the source IP for that route. This doesn't
happen if you using -I on the command-line
which means it doens't have to probe. But ping
should work without -I so add the perms.

Also grant access to network configuration
files for netutils.

dhcpc restarts chronyd
apparently without systemctl
so add a domtrans for that.