Add an example GNOME extension to disable user account button.
This is an example GNOME extension that removes the Account Settings
button in the user menu. This is mainly useful for when you remove
the user panel desktop file. The button is still present but non-functional
and confusing, so this disables it.

We stopped removing the panel by default,
but it is probably something that will be
done for production systems.

This also serves as an example and
place holder for future extensions.