Add -devel package and align with target distro
We weren't distributing some of the critical files
necessary to support development on a target running
the CLIP policy. To that end:

* add related files to a support/ directory and package them up.

* split out -devel into a separate package instead fo including
it in the base RPM to align with RH's packaging.

This also includes at least one fix to support rolling non-MCS
policies where mcs was hard-coded into the spec.