Add some "reverse interfaces" to roles for third party module support.
MOdules that want to allow staff, sysadm, and unconfined access to
their resources, such as domains and files, are can't be addressed
without modifying base policy. This is difficult for third party
modules to solve without "owning" the base policy. Add a few
convenient reverse interfaces that allow the caller to grant access
roles to their resources.