Add feature to disable building of policy SEPARATE_PKGS
If you're using Makefile.supl to specify the modules you
want in SEPARATE_PKGS, and you want to keep the build time
to a minimum and not mess with extra RPMs that are not
ultimately installed, you can us DISABLE_STD_SEPARATE_PKGS,
setting it to "y", to prevent the generated of those RPMs.

This makes maintenance a bit easier because you don't need
to modify the base CLIP Makefile to achieve this, just the
Makefile.supl.