Strip rule granting systemd logind raw block access
logind had raw block device access. Comments elsewhere
seem to indicate this might be the result of proobing for
EFI partitions for other systemd domains, but I'm not
sure why it applies here. Tested with sshd and console
logins with stripped down PAM conf and things work fine
without it.