add a systemd gneeric unit file alias for targated compat
Switching from RH's targeted policy to a pure refpol system
leads to invalidating contexts for all generic unit files. Alias
the type so it remains valid.

This solves a autorelabel bootstrap problem wherein the unit
files that support that functionality have invalid contexts
and the system freezes because those units fail.

The modifications in this commit are copyrighted works of Cubic Corporation.
(C) 2024 Cubic Corporation