alias gpt generator type in policy
Add a type alias for systemd_gpt_generator_exec_t to
systemd_generator_exec_t

This is to solve a bootstrapping problem wherein
the unit filles and generators have invalid contexts
that prevnet an initial autorelabel and reboot from
working cleanly when transitioning from targeted to
refpol.

The modifications in this commit are copyrighted works of Cubic Corporation.
(C) 2024 Cubic Corporation